This branch contains Alberto Santini's modifications.
Alberto maintained Cobcy after release 0.2, making a bunch of modifications
and two releases. There is no revision history, of course, so the splits
here are arbitrary. He rearranged the whitespace in the code quite a bit, so
it will take time to figure out what changes are actually significant. This
commit contains the deletions and moves. As you can see, many files were
deleted, compared to 0.2. I am not sure why that is. He started with an older
version 0.1, I think, and merged with 0.2 later. Lacking revision control, he
probably just copied things in until it compiled. Too bad there was no git
then :( Anyway, this commit establishes the directory structure, and the next
one will contain the actual modificaitons.